refactor: 북마크 페이지를 서버 컴포넌트로 변경
